+++ app/src/main/java/com/example/listifyjetapp/ui/screens/listItem/ListifyListItemScreen.kt
@@ -29,8 +29,8 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.listifyjetapp.R
 import com.example.listifyjetapp.widgets.texts.EmptyList
 import com.example.listifyjetapp.utils.filterListItems
-import com.example.listifyjetapp.widgets.ListifySearchBar
-import com.example.listifyjetapp.widgets.ListifyTopBar
+import com.example.listifyjetapp.widgets.bars.ListifySearchBar
+import com.example.listifyjetapp.widgets.bars.ListifyTopBar
 
 @Composable
 fun ListifyListItemScreen(
