+++ app/src/main/java/com/example/listifyjetapp/ui/screens/newList/ListifyNewListScreen.kt
@@ -17,11 +17,10 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.navigation.NavController
 import com.example.listifyjetapp.model.ListName
 import com.example.listifyjetapp.ui.screens.lists.ListsViewModel
-import com.example.listifyjetapp.widgets.FormInputField
-import com.example.listifyjetapp.widgets.ListifyTopBar
+import com.example.listifyjetapp.widgets.inputFields.FormInputField
+import com.example.listifyjetapp.widgets.bars.ListifyTopBar
 
 @Composable
 fun ListifyNewListScreen(
