+++ app/src/main/java/com/example/listifyjetapp/widgets/bars/ListifySearchBar.kt
@@ -1,4 +1,4 @@
-package com.example.listifyjetapp.widgets
+package com.example.listifyjetapp.widgets.bars
 
 
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -6,13 +6,10 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.text.input.InputTransformation.Companion.keyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.material3.TextFieldDefaults
