+++ app/src/main/java/com/example/listifyjetapp/widgets/buttons/FilledButton.kt
@@ -1,14 +1,10 @@
-package com.example.listifyjetapp.widgets
+package com.example.listifyjetapp.widgets.buttons
 
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.MailOutline
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Icon
-import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -20,35 +16,38 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
 @Composable
-fun CustomOutlinedButton(
+fun FilledButton(
     modifier: Modifier,
     shape: Shape,
-    border: BorderStroke,
+    containerColor: Color,
+    contentColor: Color,
     text: String,
-    textColor: Color,
     buttonIcon: ImageVector? = null,
-    iconDescription: String? = null,
+    iconDescription: String?,
     onClick: () -> Unit,
 ) {
-    OutlinedButton(
+    Button(
         modifier = modifier,
         shape = shape,
-        border = border,
+        colors = ButtonDefaults.buttonColors(
+            containerColor = containerColor,
+            contentColor = contentColor
+        ),
         onClick = { onClick() }
     ) {
         if (buttonIcon != null) {
             Icon(
                 imageVector = buttonIcon,
                 modifier = Modifier.size(20.dp),
-                contentDescription = iconDescription,
+                contentDescription = iconDescription
             )
         }
+
         Text(
             text = text,
-            color = textColor,
-            modifier = Modifier.padding(8.dp),
-            fontWeight = FontWeight.Bold,
+            fontWeight = FontWeight.ExtraBold,
             fontSize = 18.sp,
+            modifier = Modifier.padding(8.dp)
         )
     }
 }
